backend/pyproject.toml
[project]
name = "catchdash-backend"
version = "0.1.0"
description = "Catchdash backend API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.20",
  "httpx>=0.27.0",
  "feedparser>=6.0.11",
  "PyYAML>=6.0.2"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
